% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{create_default_par_list}
\alias{create_default_par_list}
\title{Create default parameter list}
\usage{
create_default_par_list(
  models = NULL,
  mean_fun = NULL,
  mean_class = NULL,
  err_dist = NULL,
  err_class = NULL,
  binomial_n = NA,
  method = "paired"
)
}
\arguments{
\item{models}{Models defined by set_models() object.}

\item{mean_fun}{Mean function.}

\item{mean_class}{String that indicates the class of mean functions to list.}

\item{err_dist}{Error distribution.}

\item{err_class}{String that indicates the class of error distributions to list.}

\item{binomial_n}{Value of binomial n parameter for binomial models.}

\item{method}{How mean function and error distributions should be combined;
"paired" or "crossed".}
}
\value{
List of parameter simulation objects.
}
\description{
Create parameter object using default values.
}
\examples{

## Create default parameters from set_models() object
Models <- set_models (mean_fun=c("gaussian", "beta", "sech"),
                      err_dist=c("poisson","zip"), method="crossed")
Pars <- create_default_par_list (Models)

## Create parameter object by pairing mean function and error distribution
Pars <- create_default_par_list (method="paired",
                                 mean_fun=c("gaussian", "beta"), 
                                 err_dist=c("zip", "zinb"))

## Create parameter object by crossing mean function and error distribution
Pars <- create_default_par_list (method="crossed",
                                 mean_fun=c("gaussian", "beta"), 
                                 err_dist=c("poisson", "zip", "zinb"))

}
\keyword{Simulation}
\keyword{parameters}
