% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{create_simulated_datasets}
\alias{create_simulated_datasets}
\title{Create list of simulated data sets}
\usage{
create_simulated_datasets(
  Pars,
  x = NULL,
  N = 500,
  xmin = 0,
  xmax = 100,
  seed = NULL,
  echo = FALSE
)
}
\arguments{
\item{Pars}{List containg parameters for the simulated data sets.}

\item{x}{Vector of x values to simulated data. If not given, N, xmin, and xmax must be).}

\item{N}{(optional) Sample size of data set to simulate. Default value 500.}

\item{xmin}{(optional) Minimum value of x variable. Default value 0.}

\item{xmax}{(optional) Maximum value of x variable. Default value 100.}

\item{seed}{Random number seed.}

\item{echo}{If TRUE display list describing data sets simulated}
}
\value{
Data frame containing one x variable and simulated data sets from Pars object.
}
\description{
Simulate data sets defined by the given parameter list.
}
\examples{


## Set model
Models <- set_models (mean_fun="gaussian", err_dist="negbin")

## Set list of default parameters for each model
Pars <- create_default_par_list (Models)

## Simulate data
Data <- create_simulated_datasets (Pars, seed=12345)

}
\keyword{data}
\keyword{sets}
\keyword{simulated}
