% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{error_distributions}
\alias{error_distributions}
\title{Error distributions}
\usage{
error_distributions(err_dist = NULL, err_class = NULL)
}
\arguments{
\item{err_dist}{Vector of error distribution(s).}

\item{err_class}{String that indicates the which error distributions to list.
\describe{
  \item{"binary"}{0/1 data; ("bernoulli"). }
  \item{"binomial"}{Binomial data represented as counts or proportions;
      ("binomial.count", "binomial.prop"). }
  \item{"percentage"}{[0-1] data;
      ("tab", "zitab").}
  \item{"count"}{Count data;
      ("poisson", "negbin", "zip", "zinb", "zipl", "zinbl", "zipl.mu", "zinbl.mu").}
  \item{"abundance"}{Non-negative continous data;
      ("gaussian", "tweedie", "zig", "zigl", "zigl.mu", "ziig", "ziigl", "ziigl.mu").}
}}
}
\value{
If no arguments supplied, return a data frame listing all possible error distributions
with corresponding error classes; if a vector of error classes is supplied, return
corresponding error distributions; if a vector of error distributions is supplied, return
corresponding error classes.
}
\description{
Display error distributions of given class.
}
\examples{

## Print data frame of all error distributions with corresponding error classes
error_distributions()

## Print all error distributions
error_distributions(err_class="all")

## Print error distributions for count data
error_distributions(err_class="count")

## Print error classes for poisson and gaussian data
error_distributions(err_dist=c("poisson", "gaussian"))

}
\keyword{distribution}
\keyword{error}
