% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{get_err_dist_parnames}
\alias{get_err_dist_parnames}
\title{Error distribution parameters names}
\usage{
get_err_dist_parnames(err_dist)
}
\arguments{
\item{err_dist}{Error distribution name.}
}
\value{
Vector containing parameter names of error distribution.
}
\description{
Get parameter names for given error distribution.
}
\examples{

## Get parameters for a "zinb" er
get_err_dist_parnames ("zinb")

}
\keyword{distribution}
\keyword{error}
\keyword{parameters}
