% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{get_mean_fun_parnames}
\alias{get_mean_fun_parnames}
\title{Mean function parameters names}
\usage{
get_mean_fun_parnames(mean_fun)
}
\arguments{
\item{mean_fun}{Mean function name.}
}
\value{
Vector containing parameter names of mean function.
}
\description{
Get parameter names for given mean function.
}
\examples{

## Get parameters for a "gaussian" mean function
get_mean_fun_parnames ("gaussian")

}
\keyword{function}
\keyword{mean}
\keyword{parameters}
