% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{get_parnames}
\alias{get_parnames}
\title{Model parameters}
\usage{
get_parnames(model_name)
}
\arguments{
\item{model_name}{Model_name in the form of "mean_fun-err_dist".}
}
\value{
List containing model names, character vector of mean function parameters,
error distribution parameters, and constant parameters.
}
\description{
Get parameter names for given model.
}
\examples{

## Get parameters for a "gaussian-zinb" model
get_parnames ("gaussian-zinb")

}
\keyword{model}
\keyword{parameters}
