% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{haul}
\alias{haul}
\title{Haul data set}
\format{A data frame with 1072 rows and 5 variables:
\describe{
  \item{year}{year data was collected}
  \item{x_depth}{depth that fish count was recorded, in meters}
  \item{x_lattitude}{lattitude was fish count was recorded, in degrees}
  \item{y_Albatrossia.pectoralis}{count of Albatrossia.pectoralis species}
  \item{y_Sebastolobus.altivelis}{count of Sebastolobus.altivelis}
}}
\source{
URL of data source should go here.
}
\usage{
haul
}
\description{
A data set containing the counts of two species of fish from
the NOAA Pacific Trawl Data, 2003-2004. Depth and lattitude
are also included.
}
\keyword{datasets}
