% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{haul2}
\alias{haul2}
\title{Haul data set 2}
\format{
A data frame with 1072 rows and 5 variables:
\describe{
  \item{year}{year data was collected}
  \item{depth}{depth that fish count was recorded, in meters}
  \item{lattitude}{lattitude was fish count was recorded, in degrees}
  \item{longitude}{longitude was fish count was recorded, in degrees}
  \item{Sebastolobus.altivelis}{count of Sebastolobus altivelis species}
  \item{Sebastolobus.alascanus}{count of Sebastolobus alascanus species}
  \item{Antimora.microlepis}{count of Antimora.microlepis species}
  \item{Coryphaenoides.acrolepis}{count of Coryphaenoides acrolepis species}
  \item{Alepocephalus.tenebrosus}{count of Alepocephalus tenebrosus species}
}
}
\source{
URL of data source should go here.
}
\usage{
haul2
}
\description{
A data set containing the counts of five species of fish from
the NOAA Pacific Trawl Data, 1999-2004. Depth, lattitude, and longitude
are also included.
}
\keyword{datasets}
