% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init.R
\name{init_mle}
\alias{init_mle}
\title{Initialise parameter vector for maximum likelihood algorithm.}
\usage{
init_mle(ModelInfo, Dat)
}
\arguments{
\item{ModelInfo}{Model information object from set_model_info().}

\item{Dat}{Data frame containg x and y data.}
}
\value{
Names vector containing initial parameters estimates.
}
\description{
Find initial estimates for model parameters using smoothing splines.
}
\examples{

## Simulate data and get initial model parameter estimates
Models <- set_models (mean_fun="gaussian", err_dist="zip")
ModelInfo <- set_model_info (Models[1,])
Pars   <- create_default_par_list (models=Models)
Data   <- create_simulated_datasets (Pars, seed=12345)
theta  <- init_mle (ModelInfo, Dat=Data[[1]])
print (rbind (Pars[[1]]$theta, theta))

}
\keyword{estimates}
\keyword{initial}
\keyword{parameter}
