% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{list_all_parnames}
\alias{list_all_parnames}
\title{List all model parameters}
\usage{
list_all_parnames()
}
\value{
List of all parameter names.
}
\description{
List the names of all possible model parameters.
}
\examples{

list_all_parnames()

}
\keyword{names}
\keyword{parameter}
