% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{mean_functions}
\alias{mean_functions}
\title{Mean functions}
\usage{
mean_functions(mean_fun = NULL, mean_class = NULL)
}
\arguments{
\item{mean_fun}{Vector of mean function(s).}

\item{mean_class}{String that indicates the class of mean functions to list.
Possible values are:
\describe{
  \item{"test"}{Constant and uniform mean functions.}
  \item{"main"}{Main mean functions.}
  \item{"sub"}{Special cases of main mean functions.}
  \item{"all"}{All mean functions.}
}}
}
\value{
Vector of mean functions of given class.

If no arguments are supplied, return a data frame listing all possible mean functions
with corresponding classes; if a vector of classes is supplied, return corresponding mean
functions; if a vector of mean functions is supplied, return corresponding classes.
}
\description{
Lists all possible mean functions of given class.
}
\examples{

## Print data frame of all mean functions with corresponding classes
mean_functions()

## Constant and uniform mean functions to test code
mean_functions(mean_class="test")

## Main mean functions
mean_functions(mean_class="main")

## Classes of given mean functions
mean_functions(mean_fun=c("beta", "hofIV"))

}
\keyword{function}
\keyword{mean}
