% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{msenlm.best}
\alias{msenlm.best}
\title{Find the best model for each response variable from an msenlm object}
\usage{
msenlm.best(object, best = "AICc")
}
\arguments{
\item{object}{msenlm multiple model fit.}

\item{best}{Select best model for each response variable according to
the following criteria: "nll", "AIC", "AICc", "BIC".}
}
\value{
msenlm object only containing best fitted models.
}
\description{
Find the best model for each response variable from an msenlm object
using the given criteria.
}
\examples{

\dontrun{

## Summarise data
models <- set_models (mean_fun=c("gaussian", "beta"),
                      err_dist=c("zinb", "zip"), method="crossed")
fits <- msenlm (models=models, data=haul, xvar="depth",
                yvar=c("Albatrossia.pectoralis", "Sebastolobus.altivelis"))
best <- msenlm.best (fits, best="AICc")
summary(best)
}

}
\keyword{best}
\keyword{fit}
\keyword{model}
\keyword{multiple}
\keyword{senlm}
