% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanFunctions.R
\name{mu_constant}
\alias{mu_constant}
\title{Constant mean function.}
\usage{
mu_constant(thetaM, x)
}
\arguments{
\item{thetaM}{Vector containing parameter values of mean function.}

\item{x}{Values at which to calculate the mean function.}
}
\value{
Calculates the constant mean functions defined by thetaM at points x.
}
\description{
Constant mean function.
}
\examples{

## Constant and uniform mean functions to test code
thetaM <- c(H=40)
mu_constant(thetaM, x=0:100)

}
\keyword{constant}
\keyword{function,}
\keyword{mean}
