% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanFunctions.R
\name{mu_meanfunction}
\alias{mu_meanfunction}
\title{Evaluate mean function}
\usage{
mu_meanfunction(ModelInfo, theta, x)
}
\arguments{
\item{ModelInfo}{A row from the output of set_models(), containing the
mean_function and error distribution of the model in question. Or a string
in the form of "meanfun_errdist".}

\item{theta}{Vector containing parameter values of model.}

\item{x}{Values at which to calculate the mean function.}
}
\value{
Calculates the constant mean functions defined by thetaM at points x.
}
\description{
Evalautes the mean function defined in ModelInfo, with the parameters given
in theta, at the points given by the vector x.
}
\examples{

Models <- set_models (mean_fun=c("gaussian"), err_dist=c("negbin"))
mu_meanfunction (ModelInfo=Models[1,], theta=c(H=60, m=50, s=10, phi=1.5), x=0:100)

}
\keyword{constant}
\keyword{function,}
\keyword{mean}
