% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{plot.senlm}
\alias{plot.senlm}
\title{Plot senlm model fit over data}
\usage{
\method{plot}{senlm}(x, ...)
}
\arguments{
\item{x}{Model fit.}

\item{...}{additional optional arguments.}
}
\value{
Plot of senlm model fit over data
}
\description{
Produces a scatterplot of y vs x with fitted mean curve overlaid.
}
\examples{

\dontrun{

## Plot fitted model
model <- set_models(mean_fun="gaussian", err_dist="zip")
fit <- senlm(model=model, data=haul, xvar="depth", yvar="Sebastolobus.altivelis")
plot(fit)
}

}
\keyword{fit}
\keyword{model}
\keyword{plot}
\keyword{senlm}
