% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{predict.senlm}
\alias{predict.senlm}
\title{Predict from senlm model fit}
\usage{
\method{predict}{senlm}(object, newdata, ...)
}
\arguments{
\item{object}{Model fit.}

\item{newdata}{x values to predict y values from. Optional, data values used by default.}

\item{...}{additional optional arguments.}
}
\value{
Vector of predicted y values.
}
\description{
Predict from senlm model fit using x values in newdata, or actual x values.
}
\examples{

\dontrun{

## Real data
Model <- set_models (mean_fun="gaussian", err_dist=c("zip"))
Fit <- senlm (model=Model, data=haul, xvar="x_depth", yvar="y_Sebastolobus.altivelis")
predict (Fit)
}

}
\keyword{fit}
\keyword{model}
\keyword{predict}
\keyword{senlm}
