% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{print_models}
\alias{print_models}
\title{Mean functions and error distributions}
\usage{
print_models(mean_class = "all", err_class = "all")
}
\arguments{
\item{mean_class}{String that indicates the class of mean functions to list.
Possible values are:
\describe{
  \item{"test"}{Constant and uniform mean functions.}
  \item{"main"}{Main mean functions.}
  \item{"sub"}{Special cases of main mean functions.}
  \item{"all"}{All mean functions.}
}}

\item{err_class}{String that indicates the class of error distributions to list.
\describe{
  \item{"binary"}{0/1 data; ("bernoulli"). }
  \item{"binomial"}{Binomial data represented as counts or proportions;
      ("binomial_count", "binomial_prop"). }
  \item{"percentage"}{[0-1] data;
      ("tab", "zitab").}
  \item{"count"}{Count data;
      ("poisson", "negbin", "zip", "zinb", "zipl", "zinbl", "zipl.mu", "zinbl.mu").}
  \item{"abundance"}{Non-negative continous data;
      ("gaussian", "tweedie", "zig", "zigl", "zigl.mu").}
}}
}
\value{
List containing vectors of mean functions and error distributions of given classes.
}
\description{
Display all mean functions and error distributions of given classes.
}
\examples{

## Display all mean functions and error distributions
print_models()

## Display main mean functions and count data error distributions
print_models(mean_class="main", err_class="count")

}
\keyword{distributions}
\keyword{error}
\keyword{function,}
\keyword{mean}
