% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{qres}
\alias{qres}
\title{Quantile residuals}
\usage{
qres(Fit)
}
\arguments{
\item{Fit}{senlm model fit.}
}
\value{
Quantile adjusted residuals.
}
\description{
Calculate quantiles of residuals
}
\examples{

\dontrun{
qres()
}

}
\keyword{quantile}
\keyword{residuals}
