% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{senlm}
\alias{senlm}
\title{Fit species-environment non-linear model via maximum likelihood}
\usage{
senlm(
  model = NULL,
  data = NULL,
  xvar = NULL,
  yvar = NULL,
  mean_fun = NULL,
  err_dist = NULL,
  binomial_n = NULL,
  y = NULL,
  x = NULL,
  conf.level = 0.95
)
}
\arguments{
\item{model}{Model to fit.}

\item{data}{Data frame of x and y.}

\item{xvar}{Name of x (domain) variable (or column number).}

\item{yvar}{Name of y (response) variable (or column number).}

\item{mean_fun}{Mean function (if model not supplied).}

\item{err_dist}{Error distribution (if model not supplied).}

\item{binomial_n}{The binomial n parameter, if error distribution
is "binomial.count" or "binomial.percent" (if model not supplied).}

\item{y}{Repsonse variable vector (if data not supplied).}

\item{x}{Vector of x (domain) values (if data not supplied).}

\item{conf.level}{Confidence level for parameter confidence intervals. Default is 0.95.}
}
\value{
Object containg model fit to data y and x.
}
\description{
'senlm' fits a species-environment non-linear model via maximum likelihood.
}
\examples{

\dontrun{
## Simulate data

dat  <- create_simulated_datasets(pars, N=100, xmin=0, xmax=100, seed=12345)
## Fit model
fit <- senlm(mean_fun="gaussian", err_dist="poisson", data=dat,
             xvar="x", yvar="gaussian_poisson")

## Real data
model <- set_models(mean_fun="gaussian", err_dist="zip")
fit <- senlm(model=model, data=haul, xvar="depth", yvar="Sebastolobus.altivelis")
}

}
\keyword{fit}
\keyword{mle}
\keyword{model,}
\keyword{senlm}
