% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{set_model_info}
\alias{set_model_info}
\title{Set model info}
\usage{
set_model_info(
  model = NULL,
  mean_fun = NULL,
  err_dist = NULL,
  binomial_n = NA,
  delta = NA,
  thetaC = NULL,
  data = NULL
)
}
\arguments{
\item{model}{A row from a set_models() data frame.}

\item{mean_fun}{Mean function.}

\item{err_dist}{Error distribution.}

\item{binomial_n}{Binomial size parameter, n.}

\item{delta}{Tail-adjusted beta delta parameter.}

\item{thetaC}{Alternate specification of constant parameters, either
the binomial parameter, thetaC=c(n=40);
or the tail-adjusted beta delta parameter, thetaC=c(delta=0.01).}

\item{data}{Data (x,y) set used to estimate tail-adjusted beta delta parameter.}
}
\value{
Data frame (one row) containing model information.
}
\description{
Create model information data frame (one row).
}
\examples{

Model <- set_models (mean_fun=c("gaussian"), err_dist=c("zitab"), method="crossed")
set_model_info (model=Model)
set_model_info (mean_fun="gaussian", err_dist="negbin")
set_model_info (mean_fun="gaussian", err_dist="binomial_count", binomial_n=50)
set_model_info (mean_fun="gaussian", err_dist="tab", delta=0.01)

}
\keyword{Model}
\keyword{information}
