% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{set_models}
\alias{set_models}
\title{Set models}
\usage{
set_models(
  mean_fun = NULL,
  err_dist = NULL,
  mean_class = NULL,
  err_class = NULL,
  binomial_n = NA,
  method = "crossed"
)
}
\arguments{
\item{mean_fun}{List of mean functions to create data frame of models from.}

\item{err_dist}{List of error distributions to create data frame of models from.}

\item{mean_class}{Class of mean functions to create data frame of models from (optional).}

\item{err_class}{Class of error distributions to create data frame of models from (optional).}

\item{binomial_n}{Value of binomial n parameter for binomial models (optional).}

\item{method}{Method of combining the mean functions and error distributions;
"crossed" results in all possible mean function / error distribution combinations; "paired"
matches the ith mean functions with the ith error distribution.}
}
\value{
Data frame contaning the mean functions, error distribution, and binomial n parameter
of all created models.
}
\description{
Create data frame of models to fit.
Models will be formed by either pairing or crossing the elements of
the mean function and error distribution parameters vector.
}
\examples{

## Create all possible models with supplied mean functions and error distributions
set_models(mean_fun=c("beta","sech"), err_dist=c("zip","zinb"), method="crossed")

## Combine mean functions and error distributions directly
set_models(mean_fun=c("beta","sech"), err_dist=c("zip","zinb"), method="paired")

## Create a model with a binomial n parameter
set_models(mean_fun=c("beta","sech"), err_dist=c("zip","binomial_count"), binomial_n=40)

}
\keyword{information}
\keyword{model}
