% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{set_sim_par}
\alias{set_sim_par}
\title{Set simulated model parameters}
\usage{
set_sim_par(mean_fun = NULL, err_dist = NULL, Dat = NULL,
  thetaM = NULL, thetaE = NULL, thetaC = NULL)
}
\arguments{
\item{mean_fun}{Mean function of model.}

\item{err_dist}{Error distribution of model.}

\item{Dat}{Data frame containg x and y to estimate tail-adjusted beta "delta" parameter.}

\item{thetaM}{Vector of parameters values for mean function.}

\item{thetaE}{Vector of parameters values for error distribution.}

\item{thetaC}{Vector of parameters values for constants.}
}
\value{
List containing parameters to simulated data set.
}
\description{
Set parameters for one model.
}
\examples{

## Sim simulation parameters for gaussian-negbin model
set_sim_par(mean_fun="gaussian", err_dist="negbin",
            thetaM=c(H=60,m=50,s=10), thetaE=c(phi=1.5))
## Sim simulation parameters for binomial model
set_sim_par(mean_fun="gaussian", err_dist="binomial_count",
            thetaM=c(H=60,m=50,s=10), thetaC=c(binomial_n=100))

}
\keyword{parameters}
\keyword{simulation}
