% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{simulate.senlm}
\alias{simulate.senlm}
\title{Simulate data from senlm model fit}
\usage{
\method{simulate}{senlm}(object, nsim = 1, seed = NULL,
  newdata = NULL, ...)
}
\arguments{
\item{object}{Model fit.}

\item{nsim}{Number of simulations to perform. Default value of one.}

\item{seed}{Random number seed.}

\item{newdata}{x values to predict y values from. Optional, data values used by default.}

\item{...}{additional optional arguments.}
}
\value{
Data frame of predicted y values.
}
\description{
Simulate from senlm model fit, nsim times, using x values in newdata, or actual x values.
}
\examples{

\dontrun{

## Simulate data
Model <- set_models (mean_fun="gaussian", err_dist=c("zip"))
Fit <- senlm (model=Model, data=haul, xvar="x_depth", yvar="y_Sebastolobus.altivelis")
sim(Fit, nsim=2, newdata=c(400,600,800))
}


}
\keyword{fit}
\keyword{model}
\keyword{senlm}
\keyword{simulate}
