% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{summary.msenlm}
\alias{summary.msenlm}
\title{Print summary of senlm multiple model fit}
\usage{
\method{summary}{msenlm}(object, best = NULL, ...)
}
\arguments{
\item{object}{Model fit.}

\item{best}{Select best model for each response variable according to
the following criteria: "nll", "AIC", "AICc", "BIC".}

\item{...}{Other arguments that will be ignored.}
}
\value{
Data frame summarising senlm model fits.
}
\description{
Print summary information for mulitple senlm model fits. Select the
best model fit for each response variable using the given criteria.
}
\examples{

\dontrun{

## Summarise data
models <- set_models (mean_fun=c("gaussian", "beta"),
                      err_dist=c("zinb", "zip"), method="crossed")
fits <- msenlm (models=models, data=haul, xvar="depth",
                yvar=c("Albatrossia.pectoralis", "Sebastolobus.altivelis"))
summary(fits)
summary(fits, best="AICc")
}

}
\keyword{fit}
\keyword{model}
\keyword{multiple}
\keyword{senlm}
\keyword{summary}
