% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{summary.senlm}
\alias{summary.senlm}
\title{Print summary of senlm model fit}
\usage{
\method{summary}{senlm}(object, ...)
}
\arguments{
\item{object}{Model fit.}

\item{...}{additional optional arguments.}
}
\value{
Summary of senlm model fit.
}
\description{
Print summary information for senlm model fit.
}
\examples{

\dontrun{

## Summarise data
model <- set_models(mean_fun="gaussian", err_dist="zip")
fit <- senlm(model=model, data=haul, xvar="depth", yvar="Sebastolobus.altivelis")
summary(fit)
}

}
\keyword{fit}
\keyword{model}
\keyword{senlm}
\keyword{summary}
