% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{plot.senlm}
\alias{plot.senlm}
\title{Plot senlm model fit over data}
\usage{
\method{plot}{senlm}(x, ...)
}
\arguments{
\item{x}{Model fit.}

\item{...}{additional optional arguments.}
}
\value{
Plot of senlm model fit over data
}
\description{
Produces a scatterplot of y vs x with fitted mean curve overlaid.
}
\examples{

\dontrun{

## Plot fitted model
Model <- set_models (mean_fun="gaussian", err_dist=c("zip"))
Fit <- senlm (model=Model, data=haul, xvar="x_depth", yvar="y_Sebastolobus.altivelis")
plot(Fit)
}

}
\keyword{fit}
\keyword{model}
\keyword{plot}
\keyword{senlm}
